Added parent script.
